requirements.txt
pillow
streamlit
pytest